Restante dos endpoints implementados juntamente com os Testes

+++ src/main/java/scb/microsservico/equipamentos/dto/Tranca/TrancaResponseDTO.java
@@ -6,7 +6,7 @@ import scb.microsservico.equipamentos.enums.TrancaStatus;
 @Data // Gera getters, setters e outros métodos automaticamente
 public class TrancaResponseDTO {
     private Long id;
-    private Long bicicleta;
+    private Integer bicicleta;
     private Integer numero;
     private String localizacao;
     private String anoDeFabricacao;
